Add Code generated by Creative Commons License Configurator.
File:       README.md

Use the Creative Commons license configuration page:

http://creativecommons.org/choose/

to produce HTML code that identifies my CC License, as well as
giving a link for attribution. The attribution link doesn't seem
to work when accessed from the github home page for the repo. In
debugging, I even substituted the example attribute code from the
Creative Commons web site. It didn't work either. I wasn't able
to find any reference to this phenomenon when Googling the issue,
either. So my assumption is that it's a github issue having
to do with the meta data, and that the HTML will work for Creative
Commons Attribution when accessed from another location.