Begin use of Exam Class (exam.cls) in asthma.tex.
File:       asthma.tex

The original document uses fancyhdr.sty for header formatting, but
there is a conflict between fancyhdr and exam.cls.

The following relax commands are needed to avoid conflict between
the Exam class (exam.cls) and the fancyhdr (fancyhrd.sty) package:

    \let\lhead\relax
    \let\chead\relax
    \let\rhead\relax
    \let\lfoot\relax
    \let\cfoot\relax
    \let\rfoot\relax

Now it's OK to use fancyhdr with the Exam class.